% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerile.R
\name{Dockerfile}
\alias{Dockerfile}
\title{A Dockerfile template}
\value{
A dockerfile template
}
\description{
A Dockerfile template

A Dockerfile template
}
\section{Methods}{

\describe{
  \item{\code{RUN}}{add a RUN command}
  \item{\code{ADD}}{add a ADD command}
  \item{\code{COPY}}{add a COPY command}
  \item{\code{WORKDIR}}{add a WORKDIR command}
  \item{\code{EXPOSE}}{add an EXPOSE command}
  \item{\code{VOLUME}}{add a VOLUME command}
  \item{\code{CMD}}{add a CMD command}
  \item{\code{LABEL}}{add a LABEL command}
  \item{\code{ENV}}{add a ENV command}
  \item{\code{ENTRYPOINT}}{add a ENTRYPOINT command}
  \item{\code{VOLUME}}{add a VOLUME command}
  \item{\code{USER}}{add a USER command}
  \item{\code{ARG}}{add an ARG command}
  \item{\code{ONBUILD}}{add a ONBUILD command}
  \item{\code{STOPSIGNAL}}{add a STOPSIGNAL command}
  \item{\code{HEALTHCHECK}}{add a HEALTHCHECK command}
  \item{\code{STOPSIGNAL}}{add a STOPSIGNAL command}
  \item{\code{SHELL}}{add a SHELL command}
  \item{\code{MAINTAINER}}{add a MAINTAINER command}
  \item{\code{custom}}{add a custom command}
  \item{\code{write}}{save the Dockerfile}
  \item{\code{switch_cmd}}{switch two command}
  \item{\code{remove_cmd}}{remove_cmd one or more command(s)}
}
}

\examples{
my_dock <- Dockerfile$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dockerfile$new()}}
\item \href{#method-print}{\code{Dockerfile$print()}}
\item \href{#method-RUN}{\code{Dockerfile$RUN()}}
\item \href{#method-ADD}{\code{Dockerfile$ADD()}}
\item \href{#method-COPY}{\code{Dockerfile$COPY()}}
\item \href{#method-WORKDIR}{\code{Dockerfile$WORKDIR()}}
\item \href{#method-EXPOSE}{\code{Dockerfile$EXPOSE()}}
\item \href{#method-VOLUME}{\code{Dockerfile$VOLUME()}}
\item \href{#method-CMD}{\code{Dockerfile$CMD()}}
\item \href{#method-LABEL}{\code{Dockerfile$LABEL()}}
\item \href{#method-ENV}{\code{Dockerfile$ENV()}}
\item \href{#method-ENTRYPOINT}{\code{Dockerfile$ENTRYPOINT()}}
\item \href{#method-USER}{\code{Dockerfile$USER()}}
\item \href{#method-ARG}{\code{Dockerfile$ARG()}}
\item \href{#method-ONBUILD}{\code{Dockerfile$ONBUILD()}}
\item \href{#method-STOPSIGNAL}{\code{Dockerfile$STOPSIGNAL()}}
\item \href{#method-HEALTHCHECK}{\code{Dockerfile$HEALTHCHECK()}}
\item \href{#method-SHELL}{\code{Dockerfile$SHELL()}}
\item \href{#method-MAINTAINER}{\code{Dockerfile$MAINTAINER()}}
\item \href{#method-COMMENT}{\code{Dockerfile$COMMENT()}}
\item \href{#method-custom}{\code{Dockerfile$custom()}}
\item \href{#method-read}{\code{Dockerfile$read()}}
\item \href{#method-write}{\code{Dockerfile$write()}}
\item \href{#method-merge}{\code{Dockerfile$merge()}}
\item \href{#method-remove_cmd}{\code{Dockerfile$remove_cmd()}}
\item \href{#method-remove_instr}{\code{Dockerfile$remove_instr()}}
\item \href{#method-clone}{\code{Dockerfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$new(content, file, commands, FROM = "rocker/r-base", AS)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RUN"></a>}}
\subsection{Method \code{RUN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$RUN(cmd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ADD"></a>}}
\subsection{Method \code{ADD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$ADD(from, to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-COPY"></a>}}
\subsection{Method \code{COPY()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$COPY(from, to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WORKDIR"></a>}}
\subsection{Method \code{WORKDIR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$WORKDIR(where)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EXPOSE"></a>}}
\subsection{Method \code{EXPOSE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$EXPOSE(port)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VOLUME"></a>}}
\subsection{Method \code{VOLUME()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$VOLUME(volume)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMD"></a>}}
\subsection{Method \code{CMD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$CMD(cmd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LABEL"></a>}}
\subsection{Method \code{LABEL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$LABEL(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ENV"></a>}}
\subsection{Method \code{ENV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$ENV(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ENTRYPOINT"></a>}}
\subsection{Method \code{ENTRYPOINT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$ENTRYPOINT(cmd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USER"></a>}}
\subsection{Method \code{USER()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$USER(user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARG"></a>}}
\subsection{Method \code{ARG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$ARG(arg, ahead = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ONBUILD"></a>}}
\subsection{Method \code{ONBUILD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$ONBUILD(cmd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STOPSIGNAL"></a>}}
\subsection{Method \code{STOPSIGNAL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$STOPSIGNAL(signal)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HEALTHCHECK"></a>}}
\subsection{Method \code{HEALTHCHECK()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$HEALTHCHECK(check)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SHELL"></a>}}
\subsection{Method \code{SHELL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$SHELL(shell)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAINTAINER"></a>}}
\subsection{Method \code{MAINTAINER()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$MAINTAINER(name, email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-COMMENT"></a>}}
\subsection{Method \code{COMMENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$COMMENT(comment)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-custom"></a>}}
\subsection{Method \code{custom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$custom(base, cmd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$read(Dockerfile)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$write(as = "Dockerfile")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge"></a>}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$merge(Dockerfile, remove_duplicated_comments = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_cmd"></a>}}
\subsection{Method \code{remove_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$remove_cmd(cmd, lineno)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_instr"></a>}}
\subsection{Method \code{remove_instr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$remove_instr(instr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dockerfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
