% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumper.R
\name{merge_dockerfiles}
\alias{merge_dockerfiles}
\title{merge Dockerfile}
\usage{
merge_dockerfiles(x, y)
}
\arguments{
\item{x}{string; content Dockerfile x}

\item{y}{string; content Dockerfile y}
}
\description{
merge Dockerfile
}
